Finish splitting
