+++ core/intern.go
@@ -4,8 +4,6 @@ type (
 	StringPool map[string]*string
 )
 
-var STRINGS StringPool = StringPool{}
-
 func (p StringPool) Intern(s string) *string {
 	ss, exists := p[s]
 	if exists {
