+++ core/intern_init.go
@@ -1,0 +1,5 @@
+// +build !fast_init
+
+package core
+
+var STRINGS StringPool = StringPool{}
